# [huff0] Add x86 specialisation of Decode4X