--- a/huff0/autogen.go
+++ b/huff0/autogen.go
@@ -1,0 +1,4 @@
+package huff0
+
+//go:generate go run generate.go
+//go:generate asmfmt -w decompress_amd64.s