@klauspost 2022-03-03
@WojciechMula Thanks for the great work. I will take a look and do some tests here as well.

A few pre-review notes (ignoring what is reported by tests, assuming you'll fix that)

Please make a `noasm` build tag, exclude gcc and appengine. You can grab build tags from s2: https://github.com/klauspost/compress/blob/master/s2/encodeblock_amd64.go#L3-L4

For the Go version, I found that interleaving 2 streams would give better pipelining. You removed that from the non-asm version. I will test that there isn't a regression here.

Please run [asmfmt](https://github.com/klauspost/asmfmt) on the assembly.

Is `SHRXQ` and `SHLXQ` the only vbmi used? They are (in my experience) not really faster, so a generic amd64 version should be just as fast.

Try breaking dependency chains by interleaving operations more. Your assembly is pretty much "serial", making the cpu having to work hard to re-order your code.

Zstd tests can be noisy. Checking here, literal decoding takes op 9.7% of cpu time in `DecodeAllParallel`, so I doubt you are saving 10-15%. `BenchmarkDecompress4XNoTable` is the cleanest 4X benchmark for direct comparisons.