@klauspost 2022-03-03
Numbers are looking good 👍🏼 
```
BenchmarkDecompress4XNoTable/gettysburg-32           593.83       641.27       1.08x
BenchmarkDecompress4XNoTable/twain-32                491.42       643.20       1.31x
BenchmarkDecompress4XNoTable/pngdata.001-32          718.28       829.51       1.15x
```
(these are also the only ones that have tablelog >8)

Some small regressions without asm:
```
BenchmarkDecompress4XNoTable/gettysburg-32           593.83       565.72       0.95x
BenchmarkDecompress4XNoTable/pngdata.001-32          718.28       683.05       0.95x
```