@WojciechMula 2022-03-03
> For the Go version, I found that interleaving 2 streams would give better pipelining. You removed that from the non-asm version. I will test that there isn't a regression here.

Thank you for looking at this. Yeah, I restore that freshest version. Got lost with rebasing at some point.

> Please run [asmfmt](https://github.com/klauspost/asmfmt) on the assembly.

Sure!

> Is `SHRXQ` and `SHLXQ` the only bmi used? They are (in my experience) not really faster, so a generic amd64 version should be just as fast.

Yes, I checked also SHR and SHL and the noticed BMI was faster, maybe not significantly.

> Try breaking dependency chains by interleaving operations more. Your assembly is pretty much "serial", making the cpu having to work hard to re-order your code.

Sure, will do.

> Zstd tests can be noisy. Checking here, literal decoding takes op 9.7% of cpu time in `DecodeAllParallel`, so I doubt you are saving 10-15%. `BenchmarkDecompress4XNoTable` is the cleanest 4X benchmark for direct comparisons.

Thank you for so quick response! I'll look at the regression for the plain Go version.