@klauspost 2022-03-03
Pure "amd64" speed is extremely similar, so for now bmi doesn't seem worth it:
```
BenchmarkDecompress4XNoTable/twain-32                491.42       638.31       1.30x
BenchmarkDecompress4XNoTable/pngdata.001-32          718.28       819.40       1.14x
```

You can maybe have the entire `func (d *Decoder) Decompress4X(dst, src []byte)` in an `_amd64.go`, so you can just keep the existing ask is, with opposite build tags.

My quickly thrown together code here: https://gist.github.com/klauspost/82d5c9b85c067d06d606f1c12c82615c