@WojciechMula 2022-03-04
> Without the extra "AND":
> 
> ```
> BenchmarkDecompress4XNoTable/gettysburg-32           593.83       681.84       1.15x
> BenchmarkDecompress4XNoTable/twain-32                491.42       680.16       1.38x
> BenchmarkDecompress4XNoTable/pngdata.001-32          718.28       870.23       1.21x
> ```
> 
> This also makes it "competitive" to replace that `tablelog <= 8` specialized versions, though obviously dedicated versions would likely be even better.
> 
> Only very small payloads (unlikely) are worse.

Thank you for the review. I changed almost everything that you asked for. The only big thing is to reshuffle instructions (if it's possible).

Perf results are similar on IceLake.