@klauspost 2022-03-04
> The only big thing is to reshuffle instructions (if it's possible).

You can just work on that later. The improvement is already significant.