@WojciechMula 2022-03-04
> > The only big thing is to reshuffle instructions (if it's possible).
> 
> You can just work on that later. The improvement is already significant.

@klauspost  I tried to interleave the operations as the Go code does. However, didn't notice any significant performance changes -- I'd rather say it's a noise. You can check what I did: https://github.com/WojciechMula/compress/tree/experiment. Any hints highly appreciate. :)