@klauspost 2022-03-04
Yeah, SSE <-> GPR usually is pretty slow.  Don't worry about it, let's get it working with the current improvements, we can tweak later.