@klauspost 2022-03-04
Seems the file for other platforms is missing:

`Error: huff0/decompress.go:207:26: s.Decoder().Decompress4X undefined (type *Decoder has no field or method Decompress4X)`