@WojciechMula 2022-03-07
> I have fuzz tested the current version, and it looks fine. 👍🏼 Once we get the build tags sorted, I will do a final benchmark and we can move to merging.

Great! Thank you very much for checking this. And sorry for build tags problems, TBH I'm not familiar with them. So far never used them.