@WojciechMula 2022-03-07
> > 👍🏼 Merging when tests complete.
> 
> Let me squash the commits before.

@klauspost OK, squshed the changes and added perf results to the commit message